centralize vite configurations
